+++ src/schema/schema-generator.ts
@@ -295,13 +295,15 @@ export abstract class SchemaGenerator {
                 );
                 fieldsMap[field.schemaName] = {
                   type: this.getGraphQLOutputType(field.name, field.getType(), field.typeOptions),
-                  complexity: field.complexity,
                   args: this.generateHandlerArgs(field.params!),
                   resolve: fieldResolverMetadata
                     ? createAdvancedFieldResolver(fieldResolverMetadata)
                     : createSimpleFieldResolver(field),
                   description: field.description,
                   deprecationReason: field.deprecationReason,
+                  extensions: {
+                    complexity: field.complexity,
+                  },
                 };
                 return fieldsMap;
               },
@@ -457,7 +459,9 @@ export abstract class SchemaGenerator {
         resolve: createHandlerResolver(handler),
         description: handler.description,
         deprecationReason: handler.deprecationReason,
-        complexity: handler.complexity,
+        extensions: {
+          complexity: handler.complexity,
+        },
       };
       return fields;
     }, {});