+++ tests/functional/resolvers.ts
@@ -19,7 +19,7 @@ import {
 } from "graphql";
 import * as path from "path";
 import { plainToClass } from "class-transformer";
-import { fieldConfigEstimator, simpleEstimator } from "graphql-query-complexity";
+import { fieldExtensionsEstimator, simpleEstimator } from "graphql-query-complexity";
 import ComplexityVisitor from "graphql-query-complexity/dist/QueryComplexity";
 
 import {
@@ -1112,7 +1112,7 @@ describe("Resolvers", () => {
       const context = new ValidationContext(schema, ast, typeInfo);
       const visitor = new ComplexityVisitor(context, {
         maximumComplexity,
-        estimators: [fieldConfigEstimator(), simpleEstimator({ defaultComplexity: 1 })],
+        estimators: [fieldExtensionsEstimator(), simpleEstimator({ defaultComplexity: 1 })],
       });
       visit(ast, visitWithTypeInfo(typeInfo, visitor));
       return context;
@@ -1897,7 +1897,7 @@ describe("Resolvers", () => {
       const context = new ValidationContext(schema, ast, typeInfo);
       const visitor = new ComplexityVisitor(context, {
         maximumComplexity: 2,
-        estimators: [fieldConfigEstimator(), simpleEstimator({ defaultComplexity: 1 })],
+        estimators: [fieldExtensionsEstimator(), simpleEstimator({ defaultComplexity: 1 })],
       });
       visit(ast, visitWithTypeInfo(typeInfo, visitor));
       expect(context.getErrors().length).toEqual(1);